backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor==3.3.2
pymongo==4.6.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
plotly==5.18.0
kaleido==0.2.1
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

